# BFF/src/main/resources/swagger/wardrobeSwagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /outfits/{outfitId}/users/{userId}:
    get:
      tags:
        - Outfit
      summary: API to get the user's outfit in GLAM
      description: Used to get the user's outfit
      operationId: getUserOutfit
      parameters:
        - name: outfitId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OutfitDAO'
    put:
      tags:
        - Outfit
      summary: API to update the user's outfit in GLAM
      description: Used to update the user's outfit
      operationId: updateUserOutfit
      parameters:
        - name: outfitId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutfitDAO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OutfitDAO'
    delete:
      tags:
        - Outfit
      summary: API to delete the user's outfit in GLAM
      description: Used to delete the user's outfit
      operationId: deleteUserOutfit
      parameters:
        - name: outfitId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /outfits/{outfitId}/users/{userId}/favorite:
    put:
      tags:
        - Outfit
      summary: API to set the outfit's favorite setting in GLAM
      description: Used to set the outfit's favorite setting
      operationId: updateUserOutfitFavorite
      parameters:
        - name: outfitId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: isFavorite
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OutfitDAO'
...
  /outfits/users/{userId}/save:
    post:
      tags:
        - Outfit
      summary: API to save the user outfit in GLAM
      description: Used to save the user's outfit
      operationId: saveUserOutfit
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutfitDAO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OutfitDAO'
...
  /outfits/users/{userId}:
    get:
      tags:
        - Outfit
      summary: API to get the user outfits in GLAM
      description: Used to get the user's outfits
      operationId: getUserOutfits
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutfitDAO'
components:
  schemas:
    GarmentDAO:
      type: object
      properties:
        garmentId:
          type: string
        photo:
          $ref: '#/components/schemas/GarmentPhotoDAO'
        userId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        subCategory:
          type: string
        mainColor:
          type: string
        secondColor:
          type: string
        pattern:
          type: string
        brand:
          type: string
        isFavorite:
          type: boolean
        fabric:
          type: string
        condition:
          type: string
        sleeveLength:
          type: string
        length:
          type: string
        sizeEnum:
          type: string
        patternDetail:
          type: string
        season:
          type: array
          items:
            type: string
        styles:
          type: array
          items:
            type: string
        photoUri:
          type: string
    GarmentPhotoDAO:
      type: object
      properties:
        photoId:
          type: string
        garmentId:
          type: string
        userId:
          type: string
        fileName:
          type: string
        photo:
          type: string
          format: byte
        textureUri:
          type: string
    OutfitDAO:
      type: object
      properties:
        outfitId:
          type: string
        userId:
          type: string
        name:
          type: string
        layout:
          type: string
        isFavorite:
          type: boolean
        garmentList:
          type: array
          items:
            $ref: '#/components/schemas/GarmentDAO'